refactor & clean up sprite & hud position changes
