+++ packages/frontend/src/battle-page/canvas/character-sprite/battle-character-current-sprite.tsx
@@ -1,16 +1,20 @@
-import { AnimatedComplexSpriteReact } from '@timeflies/animated-complex-sprite';
+import { AnimatedComplexSprite, AnimatedComplexSpriteReact } from '@timeflies/animated-complex-sprite';
 import { CharacterId } from '@timeflies/common';
 import React from 'react';
 import { useCharacterSpriteInfos } from './hooks/use-character-sprite-infos';
-import { useSpriteGeoUpdate } from './hooks/use-sprite-geo-update';
+import { useSpriteConfigUpdate } from './hooks/use-sprite-config-update';
+import { useSpritePositionUpdate } from './hooks/use-sprite-position-update';
 
 type BattleCharacterSprite = {
     characterId: CharacterId;
 };
 
 export const BattleCharacterCurrentSprite: React.FC<BattleCharacterSprite> = ({ characterId }) => {
+    const spriteRef = React.useRef<AnimatedComplexSprite>(null);
+
     const getPartialProps = useCharacterSpriteInfos(characterId);
-    const { spriteRef, config } = useSpriteGeoUpdate(characterId);
+    const config = useSpriteConfigUpdate(characterId);
+    const { x, y } = useSpritePositionUpdate(characterId);
 
     const partialProps = getPartialProps(config);
 
@@ -21,6 +25,7 @@ export const BattleCharacterCurrentSprite: React.FC<BattleCharacterSprite> = ({ 
     return <>
         <AnimatedComplexSpriteReact
             ref={spriteRef as any}
+            position={[ x, y ]}
             {...partialProps}
         />
     </>;
